Move time resolver to its own package
