+++ pkg/security/resolvers/time/time_resolver.go
@@ -6,7 +6,7 @@
 //go:build linux
 // +build linux
 
-package resolvers
+package time
 
 import (
 	"fmt"
@@ -16,25 +16,25 @@ import (
 	"golang.org/x/sys/unix"
 )
 
-// TimeResolver converts kernel monotonic timestamps to absolute times
-type TimeResolver struct {
+// Resolver converts kernel monotonic timestamps to absolute times
+type Resolver struct {
 	bootTime time.Time
 }
 
-// NewTimeResolver returns a new time resolver
-func NewTimeResolver() (*TimeResolver, error) {
+// NewResolver returns a new time resolver
+func NewResolver() (*Resolver, error) {
 	bt, err := host.BootTime()
 	if err != nil {
 		return nil, err
 	}
 
-	tr := TimeResolver{
+	tr := Resolver{
 		bootTime: time.Unix(int64(bt), 0),
 	}
 	return &tr, nil
 }
 
-func (tr *TimeResolver) getUptimeOffset() (time.Duration, error) {
+func (tr *Resolver) getUptimeOffset() (time.Duration, error) {
 	upTime := new(unix.Timespec)
 	err := unix.ClockGettime(unix.CLOCK_MONOTONIC, upTime)
 	if err != nil {
@@ -44,7 +44,7 @@ func (tr *TimeResolver) getUptimeOffset() (time.Duration, error) {
 }
 
 // ResolveMonotonicTimestamp converts a kernel monotonic timestamp to an absolute time
-func (tr *TimeResolver) ResolveMonotonicTimestamp(timestamp uint64) time.Time {
+func (tr *Resolver) ResolveMonotonicTimestamp(timestamp uint64) time.Time {
 	if timestamp > 0 {
 		// ignore uptime resolution failure: default back to previous behavior
 		offset, _ := tr.getUptimeOffset()
@@ -54,7 +54,7 @@ func (tr *TimeResolver) ResolveMonotonicTimestamp(timestamp uint64) time.Time {
 }
 
 // ApplyBootTime return the time re-aligned from the boot time
-func (tr *TimeResolver) ApplyBootTime(timestamp time.Time) time.Time {
+func (tr *Resolver) ApplyBootTime(timestamp time.Time) time.Time {
 	if !timestamp.IsZero() {
 		// ignore uptime resolution failure: default back to previous behavior
 		offset, _ := tr.getUptimeOffset()
@@ -64,7 +64,7 @@ func (tr *TimeResolver) ApplyBootTime(timestamp time.Time) time.Time {
 }
 
 // ComputeMonotonicTimestamp converts an absolute time to a kernel monotonic timestamp
-func (tr *TimeResolver) ComputeMonotonicTimestamp(timestamp time.Time) int64 {
+func (tr *Resolver) ComputeMonotonicTimestamp(timestamp time.Time) int64 {
 	if !timestamp.IsZero() {
 		// ignore uptime resolution failure: default back to previous behavior
 		offset, _ := tr.getUptimeOffset()
